feat(watch-history): add user watch history functionality with migration and model

+++ app/Http/Controllers/AnimeController.php
@@ -1,12 +1,13 @@
 <?php
 namespace App\Http\Controllers;
 
-use App\Http\Requests\AnimeIndexRequest;
-use App\Http\Resources\AnimeResource;
-use App\Models\Anime;
-use App\Http\Requests\AnimeThumbnailRequest;
-use App\Services\FileService;
+use Illuminate\Support\Facades\Cache;
 
+use App\Models\Anime;
+use App\Services\FileService;
+use App\Http\Requests\AnimeIndexRequest;
+use App\Http\Requests\AnimeThumbnailRequest;
+use App\Http\Resources\AnimeResource;
 
 class AnimeController extends Controller
 {
@@ -37,55 +38,67 @@ class AnimeController extends Controller
             $animes->orderBy($request->sort, 'asc');
         }
 
-        $limit  = empty($request->perPage) ? 15 : intval($request->perPage);
+        $limit = empty($request->perPage) ? 15 : intval($request->perPage);
         $offset = empty($request->page) ? 0 : (intval($request->page) - 1) * $limit;
 
         $animes = $animes->limit($limit)->offset($offset)->get();
 
         return response_success([
             'message' => 'Animes retrieved successfully',
-            'animes'  => AnimeResource::collection($animes),
+            'animes' => AnimeResource::collection($animes),
         ], 200);
     }
 
     public function show($slug)
     {
-        $anime = Anime::where('slug', $slug)->with([
-            'genres'   => function ($query) {
-                $query->select('name');
-            },
-            'seasons'  => function ($query) {
-                $query->with(['episodes']);
-            },
-            'comments' => function ($query) {
-                $query->with([
-                    'parent.parent.parent.parent.parent',
-                ])->whereNull('parent_id');
-            },
-        ])->first();
+        $seconds = 5 * 60;
+        
+        $anime = Cache::remember("anime_show_{$slug}", $seconds, function() use ($slug) {
+            return Anime::where('slug', $slug)->with([
+                'genres' => function ($query) {
+                    $query->select('name');
+                },
+                'seasons' => function ($query) {
+                    $query->with(['episodes']);
+                },
+                'comments' => function ($query) {
+                    $query->with([
+                        'parent.parent.parent.parent.parent',
+                    ])->whereNull('parent_id');
+                },
+            ])->first();
+        });
+
+        if (!$anime) {
+            return response_error('Anime not found', 404);
+        }
 
         return response_success([
             'message' => 'Anime retrieved successfully',
-            'anime'   => $anime,
+            'anime' => $anime,
         ], 200);
     }
 
     public function setAnimeThumbnail($id, AnimeThumbnailRequest $request, FileService $fileService)
     {
         $anime = Anime::find($id);
+
+        if (!$anime) {
+            return response_error('Anime not found', 404);
+        }
+
         $anime->thumbnail()->delete();
 
         $fileData = $fileService->uploadAnimeThumbnail(
-            $request->file('anime_thumbnail'),
+            $request->file('thumbnail'),
             $anime->id
         );
 
         $thumbnail = $anime->thumbnail()->create($fileData);
 
         return response_success([
-            'message' => 'Thumbnail başarıyla yüklendi',
-            'thumbnail'  => $thumbnail,
+            'message' => 'Thumbnail uploaded successfully',
+            'thumbnail' => $thumbnail
         ], 200);
     }
 }
-
